pyproject.toml
[project]
name = "success-stacker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "python-dotenv>=1.2.1",
    "werkzeug>=3.1.4",
]
